Refactored sources
