+++ pyproject.toml
@@ -10,15 +10,10 @@ authors = [
 ]
 requires-python = ">=3.10"
 dependencies = [
-    "datasets>=3.2.0",
-    "einops>=0.8.0",
-    "jsonlines>=4.0.0",
-    "nnsight>=0.3.7,<0.4.0",
     "onnx2torch>=1.5.15",
     "python-chess>=1.999",
-    "scikit-learn>=1.6.1",
     "tensordict>=0.6.2",
-    "typing-extensions>=4.12.2",
+    "torch<2.5.1",
 ]
 classifiers = [
   "Development Status :: 3 - Alpha",
@@ -46,6 +41,7 @@ backends = [
 
 [dependency-groups]
 dev = [
+    "datasets>=4.0.0",
     "gdown>=5.2.0",
     "ipykernel>=6.29.5",
     "nbconvert>=7.16.5",
@@ -53,6 +49,7 @@ dev = [
     "pre-commit>=4.0.1",
     "pytest>=8.3.4",
     "pytest-cov>=6.0.0",
+    "scikit-learn>=1.7.1",
     "v-lczero-bindings>=0.31.2",
 ]
 demo = [
