--- src/lczerolens/lenses/__init__.py
@@ -1,19 +1,0 @@
-"""
-Lenses module.
-"""
-
-from .activation import ActivationLens
-from .composite import CompositeLens
-from .gradient import GradientLens
-from .lrp import LrpLens
-from .patching import PatchingLens
-from .probing import ProbingLens
-
-__all__ = [
-    "ActivationLens",
-    "CompositeLens",
-    "GradientLens",
-    "LrpLens",
-    "PatchingLens",
-    "ProbingLens",
-]
