--- src/lczerolens/lenses/lrp/__init__.py
@@ -1,3 +1,0 @@
-from .lens import LrpLens
-
-__all__ = ["LrpLens"]
