--- src/lczerolens/lenses/probing/__init__.py
@@ -1,3 +1,0 @@
-from .lens import ProbingLens
-
-__all__ = ["ProbingLens"]
