+++ src/lczerolens/model.py
@@ -1,86 +1,126 @@
 """Class for wrapping the LCZero models."""
 
 import os
-from typing import Dict, Type, Any, Tuple, Union
+from abc import ABCMeta, abstractmethod
+from typing import Dict, Any, Union, Iterable, List, Optional, Sequence
 
 import torch
 from onnx2torch import convert
 from onnx2torch.utils.safe_shape_inference import safe_shape_inference
 from tensordict import TensorDict
 from torch import nn
-from nnsight import NNsight
-from contextlib import contextmanager
+
+from tensordict.nn import TensorDictModule
 
 from lczerolens.board import InputEncoding, LczeroBoard
 
 
-class LczeroModel(NNsight):
+class LczeroModel(TensorDictModule):
     """Class for wrapping the LCZero models."""
 
-    def trace(
+    def __init__(self, module: nn.Module, out_keys: List[str], **kwargs):
+        """
+        Parameters
+        ----------
+        module : nn.Module
+            The module to wrap.
+        out_keys : List[str]
+            The keys of the output of the module.
+        **kwargs : Any
+            Additional keyword arguments to pass to the super().__init__ method.
+
+        Raises
+        ------
+        ValueError
+            If the module is not a valid model type
+        """
+        if not isinstance(module, nn.Module):
+            raise NotImplementedError(f"Got invalid module type {type(module)}.")
+        super().__init__(module, ["boards"], out_keys, **kwargs)
+
+    def prepare_boards(
         self,
-        *inputs: Any,
-        **kwargs: Dict[str, Any],
-    ):
-        kwargs["scan"] = False
-        kwargs["validate"] = False
-        return super().trace(*inputs, **kwargs)
+        *boards: LczeroBoard,
+        input_encoding: InputEncoding = InputEncoding.INPUT_CLASSICAL_112_PLANE,
+    ) -> torch.Tensor:
+        """Prepares the boards for the model.
 
-    def _execute(self, *prepared_inputs: torch.Tensor, **kwargs) -> Any:
-        kwargs.pop("input_encoding", None)
-        kwargs.pop("input_requires_grad", None)
-        with self._ensure_proper_forward():
-            return super()._execute(*prepared_inputs, **kwargs)
+        Parameters
+        ----------
+        *boards : LczeroBoard
+            The boards to prepare.
+        input_encoding : InputEncoding, optional
+            The encoding of the boards.
 
-    def _prepare_inputs(self, *inputs: Union[LczeroBoard, torch.Tensor], **kwargs) -> Tuple[Tuple[Any], int]:
-        input_encoding = kwargs.pop("input_encoding", InputEncoding.INPUT_CLASSICAL_112_PLANE)
-        input_requires_grad = kwargs.pop("input_requires_grad", False)
-
-        if len(inputs) == 1 and isinstance(inputs[0], torch.Tensor):
-            return inputs, len(inputs[0])
-        for board in inputs:
+        Returns
+        -------
+        torch.Tensor
+            The prepared boards.
+        """
+        for board in boards:
             if not isinstance(board, LczeroBoard):
                 raise ValueError(f"Got invalid input type {type(board)}.")
 
-        tensor_list = [board.to_input_tensor(input_encoding=input_encoding).unsqueeze(0) for board in inputs]
+        tensor_list = [board.to_input_tensor(input_encoding=input_encoding).unsqueeze(0) for board in boards]
         batched_tensor = torch.cat(tensor_list, dim=0)
-        if input_requires_grad:
-            batched_tensor.requires_grad = True
         batched_tensor = batched_tensor.to(self.device)
 
-        return (batched_tensor,), len(inputs)
+        return batched_tensor
 
-    def __call__(self, *inputs, **kwargs):
-        prepared_inputs, _ = self._prepare_inputs(*inputs, **kwargs)
-        return self._execute(*prepared_inputs, **kwargs)
+    def forward(
+        self,
+        inputs: Union[TensorDict, LczeroBoard, Iterable[LczeroBoard], torch.Tensor],
+        prepare_kwargs: Optional[Dict[str, Any]] = None,
+        **kwargs,
+    ) -> TensorDict:
+        """
+        Parameters
+        ----------
+        inputs : Union[TensorDict, Iterable[LczeroBoard], torch.Tensor]
+            The inputs to the model.
+        prepare_kwargs : Optional[Dict[str, Any]], optional
+            Keyword arguments to pass to the prepare_boards method, by default None
+        **kwargs : Any
+            Additional keyword arguments to pass to the super().forward method.
 
-    def __getattr__(self, key):
-        if self._envoy._tracer is None:
-            return getattr(self._model, key)
-        return super().__getattr__(key)
+        Returns
+        -------
+        TensorDict
+            The output of the model.
+        """
+        prepare_kwargs = prepare_kwargs or {}
+        if isinstance(inputs, LczeroBoard):
+            inputs = (inputs,)
+        if not isinstance(inputs, TensorDict) and not isinstance(inputs, torch.Tensor):
+            inputs = self.prepare_boards(*inputs, **prepare_kwargs)
+        if not isinstance(inputs, TensorDict):
+            inputs = TensorDict({"boards": inputs}, batch_size=inputs.shape[0])
+        return super().forward(inputs, **kwargs)
 
-    def __setattr__(self, key, value):
-        if (
-            (key not in ("_model", "_model_key"))
-            and (isinstance(value, torch.nn.Module))
-            and (self._envoy._tracer is None)
-        ):
-            setattr(self._model, key, value)
-        else:
-            super().__setattr__(key, value)
-
-    @property
-    def device(self):
-        """Returns the device."""
-        return next(self.parameters()).device
-
-    @device.setter
-    def device(self, device: torch.device):
-        """Sets the device."""
-        self.to(device)
+    def _call_module(self, tensors: Sequence[torch.Tensor], **kwargs: Any) -> Sequence[torch.Tensor]:
+        out = super()._call_module(tensors, **kwargs)
+        return tuple(out)
 
     @classmethod
-    def from_path(cls, model_path: str) -> "LczeroModel":
+    def from_model(cls, model: nn.Module, **kwargs) -> "LczeroModel":
+        """Creates a wrapper from a model.
+
+        Parameters
+        ----------
+        model : nn.Module
+            The model to wrap.
+        **kwargs : Any
+            Additional keyword arguments to pass to the super().__init__ method.
+
+        Returns
+        -------
+        LczeroModel
+            The wrapped model instance
+        """
+        return cls(model, out_keys=cls._get_output_names(model), **kwargs)
+
+    @classmethod
+    def from_path(cls, model_path: str, **kwargs) -> "LczeroModel":
         """Creates a wrapper from a model path.
 
         Parameters
@@ -99,14 +139,14 @@ class LczeroModel(NNsight):
             If the model file extension is not supported
         """
         if model_path.endswith(".onnx"):
-            return cls.from_onnx_path(model_path)
+            return cls.from_onnx_path(model_path, **kwargs)
         elif model_path.endswith(".pt"):
-            return cls.from_torch_path(model_path)
+            return cls.from_torch_path(model_path, **kwargs)
         else:
             raise NotImplementedError(f"Model path {model_path} is not supported.")
 
     @classmethod
-    def from_onnx_path(cls, onnx_model_path: str, check: bool = True) -> "LczeroModel":
+    def from_onnx_path(cls, onnx_model_path: str, check: bool = True, **kwargs) -> "LczeroModel":
         """Builds a model from an ONNX file path.
 
         Parameters
@@ -134,12 +174,12 @@ class LczeroModel(NNsight):
             if check:
                 onnx_model = safe_shape_inference(onnx_model_path)
             onnx_torch_model = convert(onnx_model)
-            return cls(onnx_torch_model)
+            return cls.from_model(onnx_torch_model, **kwargs)
         except Exception as e:
             raise ValueError(f"Could not load model at {onnx_model_path}.") from e
 
     @classmethod
-    def from_torch_path(cls, torch_model_path: str) -> "LczeroModel":
+    def from_torch_path(cls, torch_model_path: str, **kwargs) -> "LczeroModel":
         """Builds a model from a PyTorch file path.
 
         Parameters
@@ -168,212 +208,102 @@ class LczeroModel(NNsight):
         if isinstance(torch_model, LczeroModel):
             return torch_model
         elif isinstance(torch_model, nn.Module):
-            return cls(torch_model)
+            return cls.from_model(torch_model, **kwargs)
         else:
             raise ValueError(f"Could not load model at {torch_model_path}.")
 
-    @contextmanager
-    def _ensure_proper_forward(self):
-        old_forward = self._model.forward
-
-        output_node = list(self._model.graph.nodes)[-1]
-        output_names = [n.name.replace("output_", "") for n in output_node.all_input_nodes]
-
-        def td_forward(x):
-            old_out = old_forward(x)
-            return TensorDict(
-                {name: old_out[i] for i, name in enumerate(output_names)},
-                batch_size=x.shape[0],
-            )
-
-        self._model.forward = td_forward
-        yield
-        self._model.forward = old_forward
-
-
-class Flow(LczeroModel):
-    """Base class for isolating a flow."""
-
-    _flow_type: str
-    _registry: Dict[str, Type["Flow"]] = {}
-
-    def __init__(
-        self,
-        model_key,
-        *args,
-        **kwargs,
-    ):
-        if isinstance(model_key, LczeroModel):
-            raise ValueError("Use the `from_model` classmethod to create a flow.")
-        if not self.is_compatible(model_key):
-            raise ValueError(f"The model does not have a {self._flow_type} head.")
-        super().__init__(model_key, *args, **kwargs)
-
-    @classmethod
-    def register(cls, name: str):
-        """Registers the flow.
-
-        Parameters
-        ----------
-        name : str
-            The name of the flow to register.
-
-        Returns
-        -------
-        Callable
-            Decorator function that registers the flow subclass.
-
-        Raises
-        ------
-        ValueError
-            If the flow name is already registered.
-        """
-
-        if name in cls._registry:
-            raise ValueError(f"Flow {name} already registered.")
-
-        def decorator(subclass):
-            cls._registry[name] = subclass
-            subclass._flow_type = name
-            return subclass
-
-        return decorator
-
-    @classmethod
-    def from_name(cls, name: str, *args, **kwargs) -> "Flow":
-        """Returns the flow from its name.
-
-        Parameters
-        ----------
-        name : str
-            The name of the flow to instantiate.
-        *args
-            Positional arguments passed to flow constructor.
-        **kwargs
-            Keyword arguments passed to flow constructor.
-
-        Returns
-        -------
-        Flow
-            The instantiated flow.
-
-        Raises
-        ------
-        KeyError
-            If the flow name is not found.
-        """
-        if name not in cls._registry:
-            raise KeyError(f"Flow {name} not found.")
-        return cls._registry[name](*args, **kwargs)
-
-    @classmethod
-    def from_model(cls, name: str, model: LczeroModel, *args, **kwargs) -> "Flow":
-        """Returns the flow from a model.
-
-        Parameters
-        ----------
-        name : str
-            The name of the flow to instantiate.
-        model : LczeroModel
-            The model to create the flow from.
-        *args
-            Positional arguments passed to flow constructor.
-        **kwargs
-            Keyword arguments passed to flow constructor.
-
-        Returns
-        -------
-        Flow
-            The instantiated flow.
-
-        Raises
-        ------
-        KeyError
-            If the flow name is not found.
-        """
-        if name not in cls._registry:
-            raise KeyError(f"Flow {name} not found.")
-        flow_class = cls._registry[name]
-        return flow_class(model._model, *args, **kwargs)
-
-    @classmethod
-    def is_compatible(cls, model: nn.Module) -> bool:
-        """Checks if the model is compatible with this flow.
+    @staticmethod
+    def _get_output_names(model: nn.Module) -> List[str]:
+        """Returns the output names of the model.
 
         Parameters
         ----------
         model : nn.Module
-            The model to check compatibility with.
+            The model to get the output names from.
 
         Returns
         -------
-        bool
-            Whether the model is compatible with this flow.
+        List[str]
+            The output names of the model.
         """
-        return hasattr(model, cls._flow_type) or hasattr(model, f"output/{cls._flow_type}")
-
-    @contextmanager
-    def _ensure_proper_forward(self):
-        """Rewrites the forward function to return the flow output."""
-        flow_type = getattr(self, "_flow_type", None)
-        if flow_type is None:
-            return
-
-        with super()._ensure_proper_forward():
-            old_forward = self._model.forward
-
-            def flow_forward(*inputs, **kwargs):
-                out = old_forward(*inputs, **kwargs)
-                return out[flow_type]
-
-            self._model.forward = flow_forward
-            yield
-            self._model.forward = old_forward
+        output_node = list(model.graph.nodes)[-1]
+        return [n.name.replace("output_", "") for n in output_node.all_input_nodes]
 
 
-@Flow.register("policy")
+class ForceValue(LczeroModel):
+    """Class for forcing and isolating the value flow."""
+
+    def __init__(self, module: nn.Module, out_keys: List[str], **kwargs):
+        super().__init__(module, out_keys, **kwargs)
+        output_names = self._get_output_names(self.module)
+        self._compute_value = "wdl" in output_names
+        self._wdl_index = output_names.index("wdl") if self._compute_value else None
+
+    @staticmethod
+    def _get_output_names(model: nn.Module) -> List[str]:
+        """Returns the output names of the model.
+
+        Parameters
+        ----------
+        model : nn.Module
+            The model to get the output names from.
+
+        Returns
+        -------
+        List[str]
+            The output names of the model.
+        """
+        names = LczeroModel._get_output_names(model)
+        if "value" in names:
+            return names
+        elif "wdl" in names:
+            return names + ["value"]
+        else:
+            raise ValueError("The model does not have a `value` or `wdl` head.")
+
+    def _call_module(self, tensors: Sequence[torch.Tensor], **kwargs: Any) -> Sequence[torch.Tensor]:
+        out = super()._call_module(tensors, **kwargs)
+        if self._compute_value:
+            wdl = out[self._wdl_index]
+            out = (*out, wdl @ torch.tensor([1.0, 0.0, -1.0], device=wdl.device))
+        return out
+
+
+class Flow(LczeroModel, metaclass=ABCMeta):
+    """Base class for isolating a flow."""
+
+    def __init__(self, module: nn.Module, out_keys: List[str], **kwargs):
+        if self._flow_type not in out_keys:
+            raise ValueError(f"The flow type `{self._flow_type}` is not in the output keys ({out_keys=}).")
+        filtered_out_keys = [key if key == self._flow_type else "_" for key in out_keys]
+        super().__init__(module, filtered_out_keys, **kwargs)
+
+    @property
+    @abstractmethod
+    def _flow_type(self) -> str:
+        """Returns the flow type."""
+        pass
+
+
 class PolicyFlow(Flow):
     """Class for isolating the policy flow."""
 
+    _flow_type = "policy"
 
-@Flow.register("value")
+
 class ValueFlow(Flow):
     """Class for isolating the value flow."""
 
+    _flow_type = "value"
 
-@Flow.register("wdl")
+
 class WdlFlow(Flow):
     """Class for isolating the WDL flow."""
 
+    _flow_type = "wdl"
 
-@Flow.register("mlh")
+
 class MlhFlow(Flow):
     """Class for isolating the MLH flow."""
 
-
-@Flow.register("force_value")
-class ForceValueFlow(Flow):
-    """Class for forcing and isolating the value flow."""
-
-    @classmethod
-    def is_compatible(cls, model: nn.Module):
-        return ValueFlow.is_compatible(model) or WdlFlow.is_compatible(model)
-
-    @contextmanager
-    def _ensure_proper_forward(self):
-        flow_type = getattr(self, "_flow_type", None)
-        if flow_type is None:
-            return
-
-        with LczeroModel._ensure_proper_forward(self):
-            old_forward = self._model.forward
-
-            def flow_forward(*inputs, **kwargs):
-                out = old_forward(*inputs, **kwargs)
-                if "value" in out.keys():
-                    return out["value"]
-                return out["wdl"] @ torch.tensor([1.0, 0.0, -1.0], device=out.device)
-
-            self._model.forward = flow_forward
-            yield
-            self._model.forward = old_forward
+    _flow_type = "mlh"
