--- src/lczerolens/play/__init__.py
@@ -1,10 +1,0 @@
-"""
-Play module for the lczerolens package.
-"""
-
-from .sampling import ModelSampler, Sampler, RandomSampler, PolicySampler
-from .game import Game
-from .puzzle import Puzzle
-from . import sampling, game, puzzle
-
-__all__ = ["ModelSampler", "Sampler", "RandomSampler", "PolicySampler", "Game", "Puzzle", "sampling", "game", "puzzle"]
